+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## Added
+
+* Support for reading and writing the inline and reference mask region items.
+  * Requires LibHeif 1.17.0 or later.
+
 ## v3.1.0 - 2023-05-16
 
 ## Added